+++ src/tests/smart_contract_test.ts
@@ -17,6 +17,8 @@ import {
   VALID_SMART_CONTRACT_ERC1155_MODEL,
   ERC1155_URI,
   VALID_SMART_CONTRACT_EXTERNAL_MODEL,
+  testAllReadTypesABI,
+  VALID_EXTERNAL_SMART_CONTRACT_ERC20_MODEL,
 } from "./utils";
 import { SmartContract } from "../coinbase/smart_contract";
 import { ContractEvent } from "../coinbase/contract_event";
@@ -30,8 +32,10 @@ import { TimeoutError } from "../coinbase/errors";
 
 describe("SmartContract", () => {
   let erc20Model: SmartContractModel = VALID_SMART_CONTRACT_ERC20_MODEL;
+  let erc20ExternalModel: SmartContractModel = VALID_EXTERNAL_SMART_CONTRACT_ERC20_MODEL;
   let erc721Model: SmartContractModel = VALID_SMART_CONTRACT_ERC721_MODEL;
   let erc20SmartContract: SmartContract = SmartContract.fromModel(erc20Model);
+  let erc20ExternalSmartContract: SmartContract = SmartContract.fromModel(erc20ExternalModel);
   let erc721SmartContract: SmartContract = SmartContract.fromModel(erc721Model);
   let erc1155Model: SmartContractModel = VALID_SMART_CONTRACT_ERC1155_MODEL;
   let erc1155SmartContract: SmartContract = SmartContract.fromModel(erc1155Model);
@@ -52,6 +56,118 @@ describe("SmartContract", () => {
       expect(() => SmartContract.fromModel(undefined!)).toThrow(
         "SmartContract model cannot be empty",
       );
+    });
+  });
+
+  describe(".register", () => {
+    const networkId = erc20ExternalModel.network_id;
+    const contractName = erc20ExternalModel.contract_name;
+    const contractAddress = erc20ExternalModel.contract_address;
+
+    it("should register a new smart contract", async () => {
+      Coinbase.apiClients.smartContract = smartContractApiMock;
+      Coinbase.apiClients.smartContract.registerSmartContract = jest
+        .fn()
+        .mockResolvedValue({ data: erc20ExternalModel });
+
+      const smartContract = await SmartContract.register(
+        networkId,
+        contractAddress,
+        testAllReadTypesABI,
+        contractName,
+      );
+
+      expect(Coinbase.apiClients.smartContract!.registerSmartContract).toHaveBeenCalledWith(
+        networkId,
+        contractAddress,
+        {
+          abi: JSON.stringify(testAllReadTypesABI),
+          contract_name: contractName,
+        },
+      );
+      expect(smartContract).toBeInstanceOf(SmartContract);
+      expect(smartContract.getContractAddress()).toBe(contractAddress);
+    });
+
+    it("should throw an error if register fails", async () => {
+      Coinbase.apiClients.smartContract!.registerSmartContract = jest
+        .fn()
+        .mockRejectedValue(new Error("Failed to register the smart contract"));
+      await expect(
+        SmartContract.register(networkId, contractAddress, testAllReadTypesABI, contractName),
+      ).rejects.toThrow("Failed to register the smart contract");
+    });
+  });
+
+  describe(".update", () => {
+    const networkId = erc20ExternalModel.network_id;
+    const contractAddress = erc20ExternalModel.contract_address;
+
+    const updatedContract = JSON.parse(JSON.stringify(erc20ExternalModel));
+    const updatedAbiJson = { abi: "data2" };
+    updatedContract.contract_name = "UpdatedContractName";
+    updatedContract.abi = JSON.stringify(updatedAbiJson);
+
+    it("should update an existing smart contract", async () => {
+      Coinbase.apiClients.smartContract = smartContractApiMock;
+      Coinbase.apiClients.smartContract.updateSmartContract = jest
+        .fn()
+        .mockResolvedValue({ data: updatedContract });
+
+      const smartContract = await erc20ExternalSmartContract.update(
+        updatedAbiJson,
+        updatedContract.contract_name,
+      );
+
+      expect(Coinbase.apiClients.smartContract!.updateSmartContract).toHaveBeenCalledWith(
+        networkId,
+        contractAddress,
+        {
+          abi: updatedContract.abi,
+          contract_name: updatedContract.contract_name,
+        },
+      );
+      expect(smartContract).toBeInstanceOf(SmartContract);
+      expect(smartContract.getContractAddress()).toBe(contractAddress);
+      expect(smartContract.getAbi()).toEqual(updatedAbiJson);
+      expect(smartContract.getContractName()).toEqual(updatedContract.contract_name);
+    });
+
+    it("should throw an error if update fails", async () => {
+      Coinbase.apiClients.smartContract!.updateSmartContract = jest
+        .fn()
+        .mockRejectedValue(new Error("Failed to update the smart contract"));
+      await expect(
+        erc20ExternalSmartContract.update(testAllReadTypesABI, updatedContract.contract_name),
+      ).rejects.toThrow("Failed to update the smart contract");
+    });
+  });
+
+  describe(".list", () => {
+    it("should list smart contracts", async () => {
+      Coinbase.apiClients.smartContract = smartContractApiMock;
+      Coinbase.apiClients.smartContract.listSmartContracts = jest.fn().mockResolvedValue({
+        data: {
+          data: [erc20ExternalModel],
+          has_more: true,
+          next_page: null,
+        },
+      });
+      const paginationResponse = await SmartContract.list();
+      const smartContracts = paginationResponse.data;
+
+      expect(Coinbase.apiClients.smartContract!.listSmartContracts).toHaveBeenCalledWith(undefined);
+      expect(smartContracts.length).toBe(1);
+      expect(smartContracts[0].getContractAddress()).toBe(erc20ExternalModel.contract_address);
+      expect(paginationResponse.hasMore).toBe(true);
+      expect(paginationResponse.nextPage).toBe(undefined);
+    });
+
+    it("should throw an error if list fails", async () => {
+      Coinbase.apiClients.smartContract!.listSmartContracts = mockReturnRejectedValue(
+        new APIError(""),
+      );
+      await expect(SmartContract.list()).rejects.toThrow(APIError);
     });
   });
 
@@ -333,7 +449,7 @@ describe("SmartContract", () => {
         ).rejects.toThrow(new TimeoutError("SmartContract deployment timed out"));
       });
     });
-    
+
     describe("when the contract is external", () => {
       it("throws an error", async () => {
         expect(externalSmartContract.wait()).rejects.toThrow(