+++ src/tests/utils.ts
@@ -330,6 +330,16 @@ export const VALID_SMART_CONTRACT_ERC20_MODEL: SmartContractModel = {
   },
 };
 
+export const VALID_EXTERNAL_SMART_CONTRACT_ERC20_MODEL: SmartContractModel = {
+  smart_contract_id: "test-smart-contract-1",
+  network_id: Coinbase.networks.BaseSepolia,
+  contract_name: ERC20_NAME,
+  is_external: true,
+  contract_address: "0xcontract-address",
+  type: SmartContractType.Custom,
+  abi: JSON.stringify("some-abi"),
+};
+
 export const ERC721_NAME = "Test NFT";
 export const ERC721_SYMBOL = "TEST";
 export const ERC721_BASE_URI = "https://example.com/metadata/";
@@ -781,6 +791,8 @@ export const smartContractApiMock = {
   getSmartContract: jest.fn(),
   listSmartContracts: jest.fn(),
   readContract: jest.fn(),
+  registerSmartContract: jest.fn(),
+  updateSmartContract: jest.fn(),
 };
 
 export const contractInvocationApiMock = {