+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Coinbase Node.js SDK Changelog
 
-## Unreleased
+## [0.13.0] - 2024-12-19
+
+### Added
 - Add support for registering, updating, and listing smart contracts that are
   deployed external to CDP.
 - Add support for fetching address reputation
@@ -8,6 +10,8 @@
 - Add `networkId` to `WalletData` so that it is saved with the seed data and surfaced via the export function
 - Add ability to import external wallets into CDP via a BIP-39 mnemonic phrase, as a 1-of-1 wallet
 - Add ability to import WalletData files exported by the Python CDP SDK
+
+### Deprecated
 - Deprecate `Wallet.loadSeed()` method in favor of `Wallet.loadSeedFromFile()`
 - Deprecate `Wallet.saveSeed()` method in favor of `Wallet.saveSeedToFile()`